// src/api/models/IntegrationAccountDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IntegrationAccountDto = {
  id: string;
  api_key: string;
  api_secret: string;
  email: string;
  batch_id?: string | null;
  proxy?: string | null;
  created_at: string;
};