// src/api/models/PositionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PositionDto = {
  breakEvenPrice: string;
  cumulativeFundingPayment: string;
  cumulativeInterest: string;
  entryPrice: string;
  estLiquidationPrice: string;
  imf: string;
  markPrice: string;
  mmf: string;
  netCost: string;
  netExposureNotional: string;
  netExposureQuantity: string;
  netQuantity: string;
  pnlRealized: string;
  pnlUnrealized: string;
  symbol: string;
};